+++ src/dimstack/tolerance.py
@@ -1,76 +1,5 @@
 from .utils import nround, sign_symbol
-
-
-# class Bilateral.symmetric:
-#     """Bilateral tolerancing is a method of specifying a tolerance that is symmetrical about the nominal value.
-#     This is the most common type of tolerancing.
-#     """
-
-#     def __init__(self, tol: float):
-#         self._tol = abs(tol)
-
-#     def __repr__(self) -> str:
-#         return f"Bilateral.symmetric({self._tol})"
-
-#     def __str__(self) -> str:
-#         return f"± {nround(self.T/2)}"
-
-#     @property
-#     def upper(self):
-#         return self._tol
-
-#     @property
-#     def lower(self):
-#         return -self._tol
-
-#     @property
-#     def T(self):
-#         return 2 * self._tol
-
-
-# class Bilateral.unequal:
-#     """Bilateral tolerancing is a method of specifying a tolerance that is asymmetrical about the nominal value.
-#     This can also be used for Unilateral tolerancing.
-#     """
-
-#     def __init__(self, upper: float, lower: float):
-#         if upper < lower:
-#             self.upper = lower
-#             self.lower = upper
-#         else:
-#             self.upper = upper
-#             self.lower = lower
-
-#     def __repr__(self) -> str:
-#         return f"Bilateral.unequal({self.upper}, {self.lower})"
-
-#     def __str__(self) -> str:
-#         return (
-#             f"{sign_symbol(self.upper)}{nround(abs(self.upper))} / {sign_symbol(self.lower)}{nround(abs(self.lower))}"
-#         )
-
-#     @property
-#     def T(self):
-#         return self.upper - self.lower
-
-
-# def Bilateral(upper: float, lower: float | None = None):
-#     """Automatically determine the type of bilateral tolerance to use from the upper and lower inputs.
-
-#     Args:
-#         upper (float): _description_
-#         lower (float, optional): _description_. Defaults to None.
-
-#     Returns:
-#         (Bilateral.symmetric | Bilateral.unequal): _description_
-#     """
-#     if lower is None:
-#         return Bilateral.symmetric(upper)
-#     else:
-#         if upper == lower:
-#             return Bilateral.symmetric(upper)
-#         else:
-#             return Bilateral.unequal(upper, lower)
+import numpy as np
 
 
 class Bilateral:
@@ -88,7 +17,7 @@ class Bilateral:
             self._lower = lower
 
     def __str__(self) -> str:
-        if self._upper == self._lower:
+        if np.abs(self._upper + self._lower) < np.finfo(float).eps:
             return f"± {nround(self._upper)}"
         else:
             return f"{sign_symbol(self._upper)}{nround(abs(self._upper))} / {sign_symbol(self._lower)}{nround(abs(self._lower))}"